Fix compilation with debug feature
